refactor: Removed dependencies on default maps, new types
