+++ src/colorizer/colors/categorical_palettes.ts
@@ -262,5 +262,6 @@ const keyedPaletteData: [string, PaletteData][] = rawPaletteData.map((value) => 
 });
 const paletteMap: Map<string, PaletteData> = new Map(keyedPaletteData);
 
+export type CategoricalPaletteMap = Map<string, PaletteData>;
 export const DEFAULT_CATEGORICAL_PALETTES = paletteMap;
 export const DEFAULT_CATEGORICAL_PALETTE_ID = "adobe";
