+++ src/components/ColorRampDropdown.tsx
@@ -3,8 +3,7 @@ import { Button, Tooltip } from "antd";
 import React, { ReactElement, useContext, useEffect, useMemo, useState } from "react";
 import { Color } from "three";
 
-import { ColorRamp, ColorRampType } from "../colorizer";
-import { ColorRampData, DEFAULT_CATEGORICAL_PALETTES, DEFAULT_COLOR_RAMPS } from "../constants";
+import { CategoricalPaletteMap, ColorRamp, ColorRampData, ColorRampMap, ColorRampType } from "../colorizer";
 import { AppThemeContext } from "./AppStyle";
 import styles from "./ColorRampDropdown.module.css";
 import IconButton from "./IconButton";
@@ -12,10 +11,10 @@ import IconButton from "./IconButton";
 type ColorRampSelectorProps = {
   selectedRamp: string;
   onChangeRamp: (colorRampKey: string, reversed: boolean) => void;
-  colorRamps?: typeof DEFAULT_COLOR_RAMPS;
+  colorRamps: ColorRampMap;
 
   useCategoricalPalettes?: boolean;
-  categoricalPalettes?: typeof DEFAULT_CATEGORICAL_PALETTES;
+  categoricalPalettes: CategoricalPaletteMap;
   numCategories: number;
   selectedPalette: Color[];
   onChangePalette: (newPalette: Color[]) => void;
@@ -25,10 +24,8 @@ type ColorRampSelectorProps = {
 };
 
 const defaultProps: Partial<ColorRampSelectorProps> = {
-  colorRamps: DEFAULT_COLOR_RAMPS,
   disabled: false,
   useCategoricalPalettes: false,
-  categoricalPalettes: DEFAULT_CATEGORICAL_PALETTES,
 };
 
 /**
