+++ src/constants/index.ts
@@ -1,6 +1,4 @@
 export const DEFAULT_PLAYBACK_FPS = 10;
 export const MAX_FEATURE_CATEGORIES = 12;
 
-export * from "./categorical_palettes";
-export * from "./color_ramps";
 export * from "./url";
